# swagger/swagger.yml
swagger: "2.0"
info:
  title: beego Test API
  description: beego has a very cool tools to autogenerate documents for your API
  version: 1.0.0
  termsOfService: http://beego.me/
  contact:
    email: astaxie@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
basePath: /v1
paths:
  /hero/:
    get:
      tags:
      - hero
      description: get Hero
      operationId: HeroController.Get All
      parameters:
      - in: query
        name: query
        description: Filter. e.g. col1:v1,col2:v2 ...
        type: string
      - in: query
        name: fields
        description: Fields returned. e.g. col1,col2 ...
        type: string
      - in: query
        name: sortby
        description: Sorted-by fields. e.g. col1,col2 ...
        type: string
      - in: query
        name: order
        description: Order corresponding to each sortby field, if single value, apply
          to all sortby fields. e.g. desc,asc ...
        type: string
      - in: query
        name: limit
        description: Limit the size of result set. Must be an integer
        type: string
      - in: query
        name: offset
        description: Start position of result set. Must be an integer
        type: string
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/models.Hero'
        "403":
          description: ""
    post:
      tags:
      - hero
      description: create Hero
      operationId: HeroController.Post
      parameters:
      - in: body
        name: body
        description: body for Hero content
        required: true
        schema:
          $ref: '#/definitions/models.Hero'
      responses:
        "201":
          description: '{int} models.Hero'
        "403":
          description: body is empty
  /hero/{id}:
    get:
      tags:
      - hero
      description: get Hero by ID
      operationId: HeroController.Get One
      parameters:
      - in: path
        name: id
        description: The key for staticblock
        required: true
        type: string
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/.ResponseEntity'
        "403":
          description: :id is empty
    put:
      tags:
      - hero
      description: update the Hero
      operationId: HeroController.Put
      parameters:
      - in: path
        name: id
        description: The id you want to update
        required: true
        type: string
      - in: body
        name: body
        description: body for Hero content
        required: true
        schema:
          $ref: '#/definitions/models.Hero'
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/models.Hero'
        "403":
          description: :id is not int
    delete:
      tags:
      - hero
      description: delete the Hero
      operationId: HeroController.Delete
      parameters:
      - in: path
        name: id
        description: The id you want to delete
        required: true
        type: string
      responses:
        "200":
          description: '{string} delete success!'
        "403":
          description: id is empty
  /jwt/get:
    post:
      tags:
      - jwt
      description: get JwtInfo
      operationId: JwtController.Get
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/.ResponseEntity'
        "201":
          description: body is empty
  /jwt/login:
    post:
      tags:
      - jwt
      description: Login JwtUser
      operationId: JwtController.Post
      responses:
        "200":
          description: ""
          schema:
            $ref: '#/definitions/.ResponseEntity'
        "201":
          description: body is empty
definitions:
  .ResponseEntity:
    type: object
  models.Hero:
    title: Hero
    type: object
    properties:
      create_time:
        description: 创建时间
        type: string
        format: datetime
      id:
        description: id主键
        type: integer
        format: int64
      name:
        description: 姓名
        type: string
      skills:
        description: 技能
        type: string
      type:
        description: 类型
        type: integer
        format: int64
tags:
- name: hero
  description: |
    HeroController operations for Hero
- name: jwt
  description: |
    HeroController operations for Hero